Fix use of date parsing in query API
The `parser.parse()` function will return a `datetime` object, filling
in any missing date components missing from the input string. This means
that when you call `parser.parse()` on the 29th, 30th, or 31st of a
month, and you range spans months which do not have those dates, then
those months are omitted.

The simple fix is to force the day to be the 1st, avoiding this problem.