Fix turbostat for real this time
Okay, commit f74a4c3b6 tried to fix the termination of the turbostat
command but failed to miss one critical change: the `safe_kill` call
needs the name of the process to be killed.  That previous commit
changed the behavior such that the process name is now "turbostat",
but we were still looking for "turbostat-datalog".

Along the way, fix `base-tool` to properly handle relative path
arguments which facilitate easier manual testing.