Refactor to fix trafficgen UIDs (back-port)
This is a back-port (of-sorts) of PR #2071 from `main`.  It is really a
re-implementation of that PR on the `b0.69` branch using a series of
commits to painstakingly fix up the `pbench-trafficgen` code to be a
bit more supportable and allow it to work on this branch.

WE HAVE NO GOOD WAY TO TEST THIS FUNCTIONALITY.

Not having a sufficient platform to test this functionality has proven
to make the task of maintaining this code very difficult.

The original comment from `main` is provided below.

----

The `pbench-trafficgen` script generates unique iterations based on the
following command line argument (provided or defaulted):

 * `--traffic-profiles`
 * `--rates`
 * `--traffic-directions`
 * `--frame-sizes`
 * `--num-flows`
 * `--max-loss-pcts`

However, the UIDs for each generated sample was using the following set
of data:

 * `-traffic_generator:%traffic_generator%-trial:%trial%`

This can lead to situations where one asks for `pbench-trafficgen` to
run multiple iterations, where if the trial numbers don't match the
iteration data won't be merged.

This set of modifications ensures that the UIDs used in the generated
data match the values used to create each iteration.

One of the behaviors observed previously was that one could see an
error like the following (long lines artificially wrapped):

```
Error: values do not match
key: claimed_dev_pairs values: ARRAY(0x559b88b50c20) \
  ARRAY(0x559b88b5f750)
Error: values do not match
key: null_stats values: HASH(0x559b88b5d230) HASH(0x559b88b5df68)
Error: values do not match
key: output_dir values: \
  /var/lib/pbench-agent/\
    trafficgen_RHOSP16.1-RHEL8.2-OVS-OFFLOAD\
      _tg:trex_r:none_fs:1500_nf:128_fm:si_td:bi_ml:0.002_tt:bs_\
      _2021-01-13T10:43:04/\
    1-bidirectional-1500B-128flows-0.002pct_drop/sample2 \
  /var/lib/pbench-agent/\
    trafficgen_RHOSP16.1-RHEL8.2-OVS-OFFLOAD\
      _tg:trex_r:none_fs:1500_nf:128_fm:si_td:bi_ml:0.002_tt:bs_\
      _2021-01-13T10:43:04/\
    1-bidirectional-1500B-128flows-0.002pct_drop/sample3
Error: values do not match
key: test_dev_pairs values: ARRAY(0x559b88b5d500) ARRAY(0x559b88b5f498)
```

This output was due in part to the UIDs not matching because the `trial`
value was different.  But even if we resolve that difference, the
`output_dir` key will always be different between samples, causing this
error to be emitted.  Further, differences between Perl arrays and
hashes were not being handled properly, causing the other errors
emtited above.

This change addresses those problems as well.

Further, the way values were being copied from the trial paramaeters to
the individual sample `result.json` files was correct.  Any element of
the JSON document with a 0 integer or 0.0 floating point value was
considered missing, and its value ignored. This led to missing values
for the `max_loss_pct` parameter when it was specified as 0.0. There
were other parameters values that were also restored as a result of
this fix.