Add "owner" (username) and "access" to ES documents
Extend the ES document templates with a new "phrase" defining the owning
user and "access level" for the document. The default is user "public"
and access "public" ... the ES document owner is taken from the Dataset
object describing the dataset, presumed to have been established by the
PUT operation that uploaded it to the server. (Which is not yet implemented.)