Ensure pbench develop install occurs for tests
In order for the Jenkins jobs to run correctly, we need to ensure that
the pbench python3 module is installed, including the CLI binaries that
need to be on the `PATH`.

We also correct the version of `flake8` to be the one available in
Fedora 32 (`3.7.7`), and add the missing RPMs to the development
container to avoid unnecessary pip3 installs.  We also create a
separate `docs/requirements.txt` file.

Finally, we update the branch name file to properly reference `main`
so that we use the right image tag the development container.