Enhance unit tests to capture generated commands
We had a bug in the generated `*.cmd` files for pbench-trafficgen, which
was not caught during testing. This was due in part to the fact that the
unit tests for trafficgen were failing to use the proper run directroy.

Now corrected in a previous commit, we add a dump of additional `*.cmd`
files for all test outputs, and then updated the numerous test gold
files that were affected.

Note that the `trafficgen.cmd` file in tests 22, 39, and 40 are all now
functioning correctly.

Further, we fix the `profile-builder-cmd` file to properly use the
`.cmd` suffix so that the contents of that file show up in the test
output.