Streamline server RPM spec file (#2124)
Although the Agent spec file efficiently references directories to be
packaged hierarchically, the Server spec file has a full list of the
individual packaged files. This is unnecessarily awkward to maintain,
and error prone.

The challenge here was some experimentation (the documentation
is terrible) to figure out the order of precedence across the %attr and
%defattr with directories and contained individual files. I added some
comments in the spec file to clarify.