Build state tracker (#2074)
Define a SQLAlchemy DB class "Dataset" which defines persistent state for a
Pbench dataset including most importantly the dataset's username (which is not
otherwise recorded until we index) and the dataset's current state so that we
can track the progress of a dataset through the Pbench server pipeline.

We also support Metadata associated with each Dataset, describing additional
information about datasets beyond the "state". For example, the backup
component will mark a dataset as "ARCHIVED" and pbench-reindex marks the
selected datasets to "REINDEX".

A new dataset starts in UPLOADING state and will progress through the steps as
we perform operations on it. We're defining both "in progress" -ing steps such
as INDEXING as well as completion steps (which are ready for the next
operation) such as INDEXED. There are also two "terminal" states, EXPIRED and
QUARANTINED, from which a dataset cannot exit.