Remove fake "public" user
After concerns about a reserved name in the public namespace, I set out
to remove it. The non-user is now simply None (null in Elasticsearch and
Postgresql). This ripples through many of the gold unit test files, and I
also chose to introduce real names to more of the test .db state files.

The main complication with this is that Elasticsearch doesn't index, and
won't search on, null values. After trying to re-map null values in
various ways to get around this, I realized that was still putting a
reserved string in the user namespace, and wasn't reallys solving the
problem. Instead, I've chose to re-write the query: if there's a user
name specified, we're looking for "authorization.owner" with that
string; however if no user is given, we instead look for all published
shared data by looking for "authorization.access": "public".