Make owning user required
This entails a lot of changes, and one hack:

We currently have no real user; both prep-shim and PUT are using
"pbench". This will be fixed when prep-shim goes away, and PUT
is upgraded to do real authentication.

Both the Dataset model and pbench-state-manager.py now
require that the owning user be specified. On the CLI I realized
that --user is required on --create and ignored otherwise, so I
converted it into the option value, e.g., --create=user.

I'd used a number of random user names in unit test .db files;
but now all of those users need to exist, so I pared down the
list (introducing more gold file changes than might strictly
have been necessary).

Currently, pbench-state-manager.py creates a user if the
specified user doesn't exist; this should be removed when the
real pbench-user-manager CLI tool exists, and the unittests
setup function should be converted to use it.