Add hostname and IP address validation for env
From issue #2201, it became apparent that we weren't validating
hostnames, which led to weird, but valid, directory names for
controllers being created on the server.  One way this occurs is when
the `_pbench_hostname` and or `_pbench_full_hostname` environment
variables are defined with invalid values.

We add the new commands `validate-hostname` and `validate-ipaddress`
to ensure the definition of those names are correct.

In the process of doing that, we found that we were using invalid host
names in two of our unit tests.