Userdb (#2101)
* Add "owner" (username) and "access" to ES documents

Extend the ES document templates with a new "phrase" defining the owning
user and "access level" for the document. The  document owner is taken
from the Dataset object, which is established by the PUT operation that
uploaded it to the server. (Not yet fully implemented.) The default access
is "private" until we build a mechanism to publish.

* Rename query classes

Make the resource class names consistent with the URIs for clarity; also,
although there may be some advantage in retaining consistency with the
dashboard service naming, some of the names were potentially misleading.

NOTES:

We currently have no real user; both prep-shim and PUT are using
"pbench". This will be fixed when prep-shim goes away, and PUT
is upgraded to do real authentication.

Currently, pbench-state-manager.py creates a user if the
specified user doesn't exist; this should be removed when the
real pbench-user-create CLI tool exists, and the unittests
setup function should be converted to use it.