Refactor Elasticsearch template management
This is "step 1" ... I've moved the template management class
out of indexer.py into a separate module. There is no change
in behavior, although the movement changes the Logger
path (from "<cmd>.indexer" to "<cmd>.templates") which
results in many gold file changes.

Step 2 will be to implement a Postgresql table to track the
template metadata and integrate with this class; then we can
share one source of information between the indexer and the
server APIs. I'll do step 2 as a separate PR to keep each
simple (or at least "simpler").