Enable minimum RPM version support
Fixes #2142.

We need to support a minimum version of an RPM in addition to an exact
version.  We use the `rpmdev-vercmp` from the `rpmdevtools` RPM in order
to perform the matching.

This also fixes a few of the `bench-scripts` to emit proper errors when
the required software is not installed (they were emitting debug logs
prior to this change).