Refactor PbenchTemplates
Add a new Template DB model that will allow persisting our Elasticsearch
templates to be shared between indexing and the server APIs.

The resolve() method now attempts to look up the root index name in
the templates DB table. If found, it'll check modification timestamps to
determine whether the on disk JSON (currently just resolved and
stat-ed, not read) is more recent. If not, the cached DB information is
used to create a TemplateFile object rather than reading from disk. If
the DB object doesn't exist, it's created; if it exists but is older, it's
updated.

Finally, this removes the hardcoded query index versions.

The query APIs now use the Template DB to determine the proper
full index root name with prefix and version rather than having to
hardcode.