Fix CentOS 7 / RHEL 7 pbench-agent installations
It looks like we need the latest version of `pip` to install all of
the module dependencies we have.  For all but CentOS 7 / RHEL 7 we
have the correct version.  So for those environments we now have to
install the latest version of `pip3` directly in order for us to have
our Python3 dependencies satisfied.

The days of our RHEL 7 support are numbered.

Fixes https://github.com/distributed-system-analysis/pbench/issues/2245.