Rework upload API to move filename to URI
The upload API now requires the file name, moving the controller to a
required header.  This was chosen with an eye towards the future, where
the "controller" value is not needed by the pbench server for processing.
At the time this code was written, the current pbench-agent and the
pbench-server use the controller to organize on-disk where the files are
placed.  In the future, the controller value will simply be another piece
of metadata about the file.

The API is now `/api/v1/upload/<filename>`, with the controller name
provided in the `controller` header.  This removes it from prominence in
the URI, which is why we did not move it to a URI parameter.

We also change a number of error responses to use the appropriate error
codes via `http.HTTPStatus`, and update the messages to be more succinct
while improving the context given in the logs.

Further, we fix up the `upload_api.py` module to wrap long lines, and
emit 400 response messages in the logs as warnings.  We also tighten
down some of the tests to ensure the correct log levels are being
emitted.

Finally, we bump the read & write chunk size to 64K for better through-
put.