Correct minimum RPM version support
This fixes #2289, a bug that was introduced in PR #2234 (commit
5794fc1).

This is a combination of two main problems.  The first was that we
incorrectly mocked the output of `rpm --query ...`.  It only returns
EVRs, which are version strings, which explicitly don't include the
queried RPM name.  Additionally, we were not formatting the output with
a new line, so for RPMs that have multiple versions installed, the
code would not work correctly.

The second problem was that our `pbench-sysstat` requirement was for
`12.0.3`, but the RPMs we have built publicly have version `12.0.3-1`,
which won't match exactly.

We have also corrected the `rpm` mock for the bench-scripts to behave
more closely to, but not exactly as, the actual `rpm` behavior.