Add support for `PBENCH_ORCHESTRATE` env variable
Adding support for the `PBENCH_ORCHESTRATE` environment variable is
pretty straightforward, but because the default value is not checked
against the choices, we need to separately handle verification of the
choices.

As a result of this, we ended fixing the "usage" output from the
various Tool Meister interfaces, and added a prefix for logged output.