Refactor use of `check_install_rpm`
The `check_install_rpm` was only being used by the `bench-scripts`,
and at that, they all almost did the same thing as a result of its
execution.  So instead of having layers on layers, we have combined
all the layers into one, removing much duplication.  This also has
each script record its name in the logs instead of having the name
of the common script itself.

This also removes duplicate reports regarding RPMs in the logs.