Properly collect host name information
We now use the long options to `hostname` so that the ConfigParser
behavior of lowercasing key names won't lose `-a` vs `-A`, or `-i`
vs `-I`.  And while we are at it, we refactor to make sure there is
one source of what we are collecting, adding a proper mock of the
data collected, and add the other options for data that can be
collected.