Various fixes to pbench-list-tools
Fixes #2302.

pbench-list-tools produces tracebacks in various benign
situations. This PR add a bunch of functional tests for
various scenarios (but the set is probably not exhaustive yet).

The current set of tests just checks for exit status, so it catches
all the situations where a traceback was produced, but the tests
don't check the output yet.

THere are two fixes:

- in list.py, we catch BadToolGroup exceptions (and log them).

- in base.py, we return a Path as list.py expects, not the
corresponding name.

TBD:

- return error status appropriately and change the functional tests to
expect that: currently, we always return 0.

- the tests should check the program output, at least in some cases:
in particular, calling the command with `--with-option` prints out the
options without the corresponding tool name.