Put-Shim Server
Fixes #2100

put-shim server will act as an interface between v0.69 pbench-agents
and v0.72 pbench-server. It will receive the tarballs using ssh
transfer method from 0.69 agent and will forward them to 0.72 server
using the put request.

This PR also consists of the pytests for the put-shim and the CopyResultsTb.
CopyResultTb script is used to forward the tar balls to 0.72 server using
PUT request.