list.py: enhancements to functional tests and fixes required to pass
There are two kinds of enhancements to the functional tests:

- some now require a non-zero exit status in order to pass
- the ones that require a zero exit status now check the output as
  well

The list.py execute() method now returns non-trivial status:

- 0
- 1 for bad tool groups
- 2 for anything else

This is subject to change: e.g. 2 might mean a bad setup but it
also might mean a bad --name option. We may want to split the
errors. We also probably want a symbolic representation for these
numeric codes, but that probably needs to be coordinated with changes
to the rest of the command line tools.

This commit also deals with the fall-out from the ToolCommand.tools()
method change in the previous commit which now returns a list of
paths. We convert them to names where necessary.

These changes enable all the functional tests to pass with the
exception of the ones that use `--with-option`: the current treatment
of that is inadequate and will be addressed in the third commit in
this series, which will also fix the expected output in the tests.

Finally, I let black run roughshod over the code and "fix" it: that
will allow most (but not all: see above) of the automated tests to
pass.