Publish dataset API
Implement a mechanism to change the "publication state" of a Dataset; while
mostly aimed at marking a "private" dataset to be "public" and accessible to
all, it can symmetrically be used to change the publication state back to
"private". In the future I expect it will also handle groups, allowing access
to a restricted subset of users. (TBD.)

The central complication of operations like this in Elasticsearch is that
Pbench data is spread across a large number of Elasticsearch documents which
span a number of distinct indices (run documents, TOC documents, iteration
documents, iteration sample documents, and various tool-specific documents).
Instead of searching for this set of documents dynamically using a query from
the "master" run document ID, the indexer builds a document map recording all
indices and document IDs representing the dataset.