Rename pbench-agent.cfg.example file
Fixes #2343,

We rename the example config file to pbench-agent.cfg. That allows
the basic operations of the pbench-agent to succeed without further
steps other than installing the RPM. However, moving results to the
server will not succeed, until a proper config file and ssh key file
are installed, so we still expect people to install using Ansible.