Expect `trafficgen` to be installed already
Benchmark wrappers should not merge the concern of how software is to
be installed and deployed with how to run that software.  Those are two
separate concerns that create a tightly coupled set of behaviors where
the benchmark wrapper has to "know" how the software should be built
and installed.

For `trafficgen`, we now expect the code to be installed in the expected
location, `/opt/bench-trafficgen/trafficgen`, with the installation
process providing the version number of the software in the `VERSION`
file.