Quick-n-Dirty bit-rot detection for tar balls
This is a quick capture of the current bit-rot detection that has been
deployed as a stop-gap whilest backup-verification is off.

It has a bug in that any empty or missing `.md5` files for `.tar.xz`
files are not detected.

This commit captures the code used in the live deployment and does not
attempt to fix any problems found there-in.