Make the default tool set lighter
Fixes #1714.
Fixes  #2331.
Related to #1679.
Related to #2510.

Define three tool set levels (`light`, `medium`, and `heavy`) and redefine the `default` to be the `medium` one, avoiding `pidstat` and `perf`.

The three new tool sets are:

 - `light` -- `vmstat` (providing CPU, memory and IO stats)
 - `medium` -- the light tool set plus `iostat` and `sar` (more detailed CPU/mem/IO stats, plus network stats)
 - `heavy` -- the medium tool set plus `mpstat`, `perf`, `pidstat`, `proc-interrupts`, `proc-vmstat`, `turbostat` - the "kitchen sink".

In addition, the old default tool set is preserved in a `legacy` tool set, allowing users who want that behavior to get it relatively painlessly.

Putting the new definitions into the `[DEFAULT]` section allows overriding the `default` tool set in the `pbench-agent.cfg` file without having to touch `pbench-agent-default.cfg`.

Use the `--toolset` option of `pbench-register-tool-set` to choose a non-`default` tool set:
```
pbench-register-tool-set --toolset=legacy ....
```
or override the `default` tool set in the `pbench-agent.cfg`:
```
[pbench/tools]
...
default-tool-set = %(legacy-tool-set)s
...
```

----
The redefinition of the `default` tool set will require publicity to make the user base aware of the change, as well as of what to do to get back the old set.

See discussion in https://github.com/distributed-system-analysis/pbench/discussions/2510.