Fio (et al.) postprocessing performance improvements
This change modifies the support in BenchPostprocess.pm to expect
timeseries data in the workload hash to be a hash of hashes, using the
timestamp as the key instead of an array of hashes.  This ensures that
each timestamp is unique, and it makes each timestamp's data accessible
in constant-time rather than requiring a linear (O(n^2)) search.  For
large datasets, this results in a dramatic improvement in execution time
for producing the aggregate results.

Also:
* Do not re-write fio job file in post-process only mode
* Address code formatting and other code quality issues
* Tweak agent/tool-scripts/postprocess unit test script
  * Limit the number of lines in result.json for fio tests
  * Remove the reference-result symlink from the process-iteration-samples-0 gold
* Add unit test for post-process only with no directory
* Add a test which stresses fio-postprocess performance

Co-authored-by: Peter Portante <peter.portante@redhat.com>