Correct `tracker` module name
When I started the "dataset tracker" I named the module `tracker` and
the class `Dataset` and somehow I never thought to make them match
before submitting. This is increasingly bugging me, so I'm taking a
moment to fix that oversight and make them consistent.

These changes should be very easy to review.