Add dataset Metadata needed for Overview Page
This enhances the `PUT` to instantiate some dataset metadata to
enable tracking whether a dataset is SEEN, SAVED, PUBLISHED, and
the expiration (DELETION) date; and enhances the datasets/list and
datasets/detail response payload with a new `"serverMetadata"`
section that allows a client to access that data.

The promotion of the Metadata value column to JSON (long overdue)
resulted in massive changes to the gold files because, for some reason
(and even though I believe sqlite3 actually represents JSON columns as
TEXT), there were encoding changes that, ironically, didn't affect the
text but radically changed the value of sqlite3's `length` function on
the column. I gave up and removed the truncation, so that the gold
files include the full document map.