Switch to --user PIP installation
All pbench dependencies (not already installed via RPM) will be put
in the user pbench site packages. A pbench.pth file is created in the
pbench site package directory to cause the site package (which is
implicitly loaded when Python starts) to also add the installation lib
directory to the path.

The result is that when logged in as pbench, you will be able to run
the interactive Pbench commands without setting PYTHONPATH
explicitly.

If you really want to run Pbench interactive commands as root, you can,
for example, use

su pbench -c "/opt/pbench-server/bin/pbench-user-list"

to get the benefit of the implicit pbench account configuration.

This also adds creation of the pbench user and group to the RPM
spec file to be sure they exist. (The %post already depended on
them existing, but they were actually created after RPM installation
via Ansible, and this PR increases the dependency.)