Check for the `ansible` RPM for the `pprof` tool
The `pprof` tool requires both `golang` and `ansible` RPMs, but the tool
was only checking for `golang`.  We add an explicit check for both now.

In principle the `pbench-agent` should not have any explicit RPM
dependencies for RPMs needed by tools.  But, the `pbench-run-bencmark`
interface requires the `ansible` RPM, so we happen to satisfy the
`pprof` requirement indirectly.

However, we don't want to have the `ansible` Python3 module explicitly
listed in the `agent/requirements.txt` file since it is not a direct
dependency for any Python3 code of the agent.  So for testing purposes
we move the `ansible` module requirement to `test-requirements.txt`.