Refactor `results-move` and bring to parity
We significantly refactor the implementation of the
`pbench-results-move` command so that it matches the expected behaviors
set by the legacy `pbench-move-results`.  This included ensure the tar
ball was created with the correct directory structure, adding
environmental metadata, and correcting the controller name
appropriately.

We drop prompting for `--token`, we no longer require `--controller`,
defaulting to the current host name.  We clean up the temporary
directory, correct type hints, add doc strings, use existing agent
logger, and emit status to stdout.

We also:
 * Move push to its own module
 * Rename `results_push` to `main`
 * Drop explicit `pbench-agent` logger from push
 * Refactor `move` results to own module
 * Provide `pbench-results-move` in the RPM
 * Add the new `--copy-only` switch
 * Rename `test_move_results.py` to `test_results_move.py`
 * Rename `--copy` to `--delete/--no-delete`
 * Update command line in help text of results cmds