Correct checking for required tool packages
We had hard-coded the names of all the tools that needed to have their
package checks performed, but failed to keep that list updated correctly.

Instead, we use the `tool_package_name` variable existence to determine
when to check for a package, and only the two special cases, `pprof` and
`systemtap`, have added logic.