More installation cleanups (#2452)
This is a small incremental change towards issue #2422.

Fix the pbench-server service file to set up a proper PATH to find a pbench --user installed gunicorn binary

The previous Environment statement in the pbench-server.service file wasn't quite doing what we thought it was, because these are not executed in a context where the $PYTHONPATH is substituted. (It's also not necessary since #2423.) This changes ExecStart to run bash and construct a proper PATH that can resolve a locally installed gunicorn so the server will start. (In the longer run, we should be installing the python-gunicorn RPM instead of resolving this from pip install.)

Note that the proper PATH won't change after installation, however the cost of the bash is small and this is a common systemd service pattern. The alternative would be to generate pbench-server.service at installation time from a template, using information gleaned from the pbench account (basically PATH=$(su pbench -c "echo ${PATH}:\$(python3 -m site --user-base)/bin"))