Implementation of #2370 query semantics
This allows searching for combinations of "user" and "access"
parameters in Elasticsearch-based queries with much more
flexibility than previously. For example, to find all "private"
datasets owned by a specific user.

The docstring for the new query builder documents all of the
odd cases. The big issue is that there are conflicts between the
intended syntax here and the authorization checker: for example
while {"access": "private"} for an authenticated user should return
the user's private datasets, right now I'm being told that the user
doesn't have access to user "None"'s data, and I need to fix that;
although I'd intended to clean up authorization messaging in a
separate PR, the streams may be crossing here.

I added a pure unit test to examine the various query builder
cases, but I'm still doing real-world testing on a deployed VM
and as mentioned above, not everything is working as it should.
Never the less, I'm feeling "close enough" to push a draft PR both
to archive what I have and to let the team see it (should they be
so inclined... Webb, I'm virtually looking at you).