pbench-list-tools: add tests and fixes for exceptions
Add some basic functional tests. The last three are failing tests,
designed to demonstrate the problems described in issues #2302, #2345
and #2346.

-----
Catch BadToolGroup exception and return error

Fixes #2302

Running `pbench-init-tools -g foo', where `foo' is an unknown group,
results in an exception.

Modify the ListTools().execute() function to catch the exception,
log an error and return an error status of 0, the 0.69.9 behavior.

-----
Fix type confusion that led to an uncaught exception

Fixes #2346

`pbench-list-tools -o' got an exception because a tool name (a string)
was used as a Path, trying to read the file that contains the options
for the tool. We convert the tool name into a Path in order to read
the file.

The output of the command was just the options without the name of
the associated tool, so we now store the name of the tool as well.
The output contains the same information as the 0.69.9 version, but
the format is different.

the 0.69 version produces output like this:
```
default: iostat --interval=3,mpstat --interval=3,perf
--record-opts='record -a --freq=100', ...
```
whereas the current version's output is straight print() output
from python:

```
default: alphaville.usersys.redhat.com [('iostat',
'--interval=3\n'), ('mpstat', '--interval=3\n'), ('perf',
"--record-opts='-a --freq=100'\n"), ...
```

The format is ugly but will be fixed in a subsequent PR:
this PR is only dealing with the tracebacks identified
in the associated issues.

-----
Skip files when looking for hosts under a tool group

Fixes #2345

When listing tools for a given host under a given tool group,
skip non-directory files (e.g. __label__ and __trigger__).