Remove "authorization" from convert_username
Leave authorization to the authorization logic. Note, there's one
hole here I'll address in a separate commit: there's currently a
difference in behavior when an unauthenticated client attempts to
query data for a user that exists (FORBIDDEN) vs a user that
doesn't exist (BAD_REQUEST), and I'm thinking about the best
way to close that. (I'm considering making it an error, at
schema validation time, for an unauthenticated client to include
the `"user"` parameter. An UNAUTHORIZED return would be
appropriate.)