Remove authentication from username validation
`convert_username` no longer checks for authorization of an
authenticated user to the specified username, leaving that to
`_check_authorization()`. However it does check whether the
client authenticates; unauthenticated clients are not to allowed
to see whether a specified username is valid, thus all "user"
parameter references from an unauthenticated client are
judged violations (UNAUTHORIZED / 401).

In extension of this, separate the current muddied permission
violations into clearer FORBIDDEN / 403 (the client *is*
authenticated, but lacks permission to access the specified
resource) and UNAUTHORIZED / 401 (the client is not
authenticated, and authentication is required). This distinction
allows a client (e.g., web UI) to redirect to a login page on
401 errors while giving an error on 403 errors, as intended.

Along the way I consolidated the test logic to determine an
expected failure status into a common _get_expected_status
method for greater consistency (although there are a few odd
variations).