Add dataset Metadata needed for Overview Page (#2417)
Move the Schema classes to a new ApiBase in server.api.resources, and
reworked the post/put/get delegation to (hopefully) support more general
interfaces. I use GET and PUT for metadata management.

The promotion of the Metadata value column to JSON (long overdue)
resulted in massive changes to the gold files because the change
resulted in encoding changes affecting the reported column `length`
result even though the characters are the same.

Metadata can be requested by the `/datasets/list` and `/datasets/detail`
APIs, as well as by a GET on the new `/datasets/metadata` API. The latter
takes all parameters as query strings on the URI.

The code supports four "namespaces". The "dataset" namespace exposes
some columns of the primary dataset DB table, owner and access. The "system"
namespace exposes Pbench-owned metadata such as "deletion" (the dataset
expiration timestamp). The "dashboard" and "user" namespaces are reserved
for free external client use, with any acceptable JSON key hierarch in a dotted
path structure, such as "dashboard.saved" or "dashboard.seen", or
"user.contact.email".