Create load_pbench_runs() method
The pbench run data is much smaller than the result data, so we
create a very simple routine to load all the run data, apply the
proper filters, massage it a bit, and store only the fields we
need when processing the result data.  This provides a much smaller
memory footprint.