Use a generator pattern for processing results
We use a generator to load one record at a time from the JSON
docs loaded from a file, process it, adorn it with pbench run
data and then write it out without storing it.  This avoids an
using memory to store all the data while processing.

This does not implement the client names and/or disk/host names
fetching required to complete the data adornment.